Extract NotificationBroadcastReceiverHandler to its own file
